chart issue for mouse accel
so i recently been playing around with my settings in mouse accel and my app just crashed and i try to re open it and it says invalid chart error. I tried to delete the program and re download it but it still does not work as if my computer blocked the program from being used. It says "invalid chart range error" and has whole bunch of numbers. If someone can help me with this id greatly appreciate it!!!